Improve latest episode of podcast on homepage.
This changes the layout to cover the full width, mentions the
word "Episode" to clarify the role of the number, and also adds
a header indication we're talking about the podcast at all.

I refactored some of the "design" for section headers into a
partial / template. Much more can be done here, but I did not
want to entangle major redesign with this otherwise simple
incremental change.